Add Github action to set up snowcli in a CI/CD workflow  (#1)
This action installs snowflake-cli-labs in an isolated environment using
pipx, so it will not have conflicts with any other package. In addition
there is an optional parameter to set up a default config file in
.snowflake directory.

There is not a conventional way to test a composite github action
because there is not a framework for that, instead some
usage examples were added to ensure a valid correctness and utility of
the github action. A brief description of every example jobs and what
they do are the following:

**install_Snowcli_with_pip_and_then_with_pipx:** This job installs
snowcli version 2.1.0 with pipx and version 2.1.1 with pip, the workflow
should show 2.1.0 when executes snow --version.

**install_in_multiple_os:** This job validates the correct installation
of snowcli in any OS.

**deploy_demo_native_app:** This is a more complex that set up snowflake
cli and creates a version for the applicaiton package defined in the
source code.

---------

Co-authored-by: Tomasz Urbaszek <tomasz.urbaszek@snowflake.com>